add validation to rest api

+++ src/main/java/home/work/filmolike/controller/NoteRestController.java
@@ -2,65 +2,70 @@ package home.work.filmolike.controller;
 
 import home.work.filmolike.domain.Note;
 import home.work.filmolike.domain.User;
+import home.work.filmolike.dto.NoteConverter;
 import home.work.filmolike.dto.NoteDto;
 import home.work.filmolike.service.NoteService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
+import javax.validation.constraints.Min;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.List;
 import java.util.Optional;
 
+//Использую DTO осознанно, чтобы показать, что имею о них знания.
+//Несмотря на то, что такое решение приводит к дублированию кода
+
 @RestController
+@Validated
 @RequestMapping("/rest")
 public class NoteRestController {
     private static final Logger LOGGER = LoggerFactory.getLogger(NoteRestController.class);
 
     private final NoteService noteService;
 
-    @Autowired
     public NoteRestController(NoteService noteService) {
         this.noteService = noteService;
     }
 
     @GetMapping("/note/{id}")
-    public ResponseEntity<?> getNoteById(@PathVariable Long id) {
-        NoteDto noteDto = noteService.findByIdDto(id);
+    public ResponseEntity<?> getNoteById(
+            @PathVariable @Min(1) Long id,
+            @AuthenticationPrincipal User user)
+    {
+        Optional<Note> noteOptional = noteService.findById(id);
 
-        if (noteDto.equals(NoteDto.NULL_NOTE)) {
+        if (!noteOptional.isPresent()) {
             return ResponseEntity.notFound().build();
         }
+
+        Note foundNote = noteOptional.get();
+
+        if (hasNotRights(foundNote, user)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Нет прав для просмотра данных");
+        }
+
+//        NoteDto noteDto = noteService.findByIdDto(id);
+
+//        if (noteDto.equals(NoteDto.NULL_NOTE)) {
+//            return ResponseEntity.notFound().build();
+//        }
 
         try {
             return ResponseEntity
                     .ok()
-                    .location(new URI("/rest/note/" + noteDto.getId()))
-                    .body(noteDto);
+                    .location(new URI("/rest/note/" + foundNote.getId()))
+                    .body(NoteConverter.toDto(foundNote));
         } catch (URISyntaxException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
         }
-
-//        return noteService.findById(id)
-//                .map(note -> {
-//                    try {
-//                        NoteDto noteDto = NoteDto.toDto(note);
-//
-//                        return ResponseEntity
-//                                .ok()
-//                                .location(new URI("/rest/note/" + noteDto.getId()))
-//                                .body(noteDto);
-//                    } catch (URISyntaxException e) {
-//                        return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
-//                    }
-//                })
-//                .orElse(ResponseEntity.notFound().build());
     }
 
     @GetMapping("/notes")
@@ -78,51 +83,77 @@ public class NoteRestController {
 
     @PostMapping("/note")
     public ResponseEntity<NoteDto> createNote(
-            @RequestBody Note note,
+            @Valid @RequestBody NoteDto noteDto,
             @AuthenticationPrincipal User user
     ) {
-        LOGGER.info("Received note: title: {}, watched: {}, estimate: {}", note.getTitle(), note.isWatched(), note.getEstimate());
+        LOGGER.info("Received note: title: {}, watched: {}, estimate: {}", noteDto.getTitle(), noteDto.getWatched(), noteDto.getEstimate());
 
-        Note savedNote = noteService.save(note, user);
+        Note savedNote = noteService.save(NoteConverter.toDomain(noteDto), user);
+
+        LOGGER.info("Saved note: id: {}, title: {}, watched: {}, estimate: {}", savedNote.getId(), savedNote.getTitle(), savedNote.isWatched(), savedNote.getEstimate());
 
         try {
             return ResponseEntity
                     .created(new URI("/rest/note/" + savedNote.getId()))
-                    .body(NoteDto.toDto(savedNote));
+                    .body(NoteConverter.toDto(savedNote));
         } catch (URISyntaxException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
         }
     }
 
     @PutMapping("/note/{id}")
-    public ResponseEntity<NoteDto> updateNote(
+    public ResponseEntity<?> updateNote(
             @PathVariable Long id,
-            @RequestBody Note note,
+            @Valid @RequestBody NoteDto noteDto,
             @AuthenticationPrincipal User user
     ) {
-        Optional<Note> existingNote = noteService.findById(id);
+        Optional<Note> noteOptional = noteService.findById(id);
 
-        if (!existingNote.isPresent()) {
+        if (!noteOptional.isPresent()) {
             return ResponseEntity.notFound().build();
         }
 
-        note.setId(id);
+        Note foundNote = noteOptional.get();
 
-        Note updatedNote = noteService.save(note, user);
+        if (hasNotRights(foundNote, user)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Нет прав для изменения данных");
+        }
+
+        Note updatingNote = NoteConverter.toDomain(noteDto);
+        updatingNote.setId(id);
+
+        Note updatedNote = noteService.save(updatingNote, user);
 
         try {
             return ResponseEntity
                     .ok()
-                    .location(new URI("/rest/note/" + note.getId()))
-                    .body(NoteDto.toDto(updatedNote));
+                    .location(new URI("/rest/note/" + updatedNote.getId()))
+                    .body(NoteConverter.toDto(updatedNote));
         } catch (URISyntaxException e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).build();
         }
     }
 
     @DeleteMapping("/note/{id}")
-    public ResponseEntity deleteNote(@PathVariable Long id) {
+    public ResponseEntity<?> deleteNote(
+            @PathVariable Long id,
+            @AuthenticationPrincipal User user)
+    {
+        Optional<Note> noteOptional = noteService.findById(id);
+
+        if (!noteOptional.isPresent()) {
+            return ResponseEntity.notFound().build();
+        }
+
+        if (hasNotRights(noteOptional.get(), user)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body("Нет прав для удаления данных");
+        }
+
         noteService.delete(id);
         return ResponseEntity.ok().build();
+    }
+
+    private boolean hasNotRights(Note note, User user) {
+        return !user.equals(note.getUser());
     }
 }
